Cont
Added a new button to mess around with placing a compartment.
Gave it logic and toooo many methods. Some of this will have to be moved to the Build controller im making.
The placement shadow deals with raycasting. It continuously casts rays to give a visual representation of where the player can and can not build. At build it also fires one last raycast from all cannons and gives a positive/negative on being allowed to build there. It also contains some placement logic functions.