Sprite test
Testing on how loading in a sprite for a compartment could work.
Passing the sprites trough the editor is annoying because I have to use a list, not a dictionary ugh.
Added a third child object to the cubes that make up a column. This will hold the sprite.
The placement grid should probably be a lighter color and disabled outside build mode.
I have to start with the merging of compartments (tiers) tomorrow. That is one of the major things missing and if I continue without it I will have to redo a lot of logic later
-XOXO RV