Tiers v1
Continuing the implementation of the building system.
Compartments are made up of an object that has both the CombinedCompartment script and CompartmentType script.
Each (non empty type) CombinedCompartment script having object has multiple "subrooms" as children, these do not have a script themselves. Each time a compartment is merged into an existing CombinedCompartment of the same type, a new "subroom" is created. The amount subrooms a CombinedCompartment has corresponds to the tier of the compartment. A subroom has the corresponding columns as children.
The subrooms exists to make things like downgrading tiers and undoing building actions easier. The compartment logic should be housed in the CombinedCompartment (or ship), not there.