Major overhaul of placement shadow
NB! The shadow is only visual at the moment. If it shows red you can still probably place the compartment.
I had to rework the entire shadow code.
Before the individual squares turned red when either
a) it was over another compartment
b) it was not over the compartment space.
Now it (to some extent) follows the rules of building.
1) It checks a and b
2) it checks (the shadows) each row, and checks if any of the squares is next to any column containing a non empty type compartment. If at least one of the squares in a row is next to one, the entire row is turned green.
Well, turned green is the wrong way around. The whole thing is calculated with two matrixes. The placement shadows starts green and step by step tiles that fail the checks are turned red.